refactor: lib 리팩토링
